(chore) Move building components to their right folder.
