+++ src/app/intervention-building/contact/contact.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { TestModule } from '../../test.module';
 import { ContactComponent } from './contact.component';
-import { BuildingContactService } from '../shared/services/building-contact.service';
+import { BuildingContactService } from '../../intervention-survey/shared/services/building-contact.service';
 
 describe('ContactComponent', () => {
   let component: ContactComponent;
