+++ src/app/intervention-building/dangerous-material/dangerous-material.component.html
