+++ src/app/intervention-building/dangerous-material/dangerous-material.component.spec.ts
@@ -2,9 +2,9 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { TestModule } from '../../test.module';
 import { DangerousMaterialComponent } from './dangerous-material.component';
-import {BuildingHazardousMaterialService} from '../shared/services/building-hazardous-material.service';
-import {HazardousMaterialService} from '../shared/services/hazardous-material.service';
-import {UnitOfMeasureService} from '../shared/services/unit-of-measure.service';
+import {BuildingHazardousMaterialService} from '../../intervention-survey/shared/services/building-hazardous-material.service';
+import {HazardousMaterialService} from '../../intervention-survey/shared/services/hazardous-material.service';
+import {UnitOfMeasureService} from '../../intervention-survey/shared/services/unit-of-measure.service';
 
 describe('DangerousMaterialComponent', () => {
   let component: DangerousMaterialComponent;
