+++ src/app/intervention-building/person-requiring-assistance/person-requiring-assistance.component.styl
