+++ src/app/intervention-survey/intervention-survey.component.spec.ts
@@ -6,13 +6,13 @@ import { InterventionSurveyComponent } from './intervention-survey.component';
 import { BuildingComponent } from './building/building.component';
 import { BuildingDetailComponent } from './building-detail/building-detail.component';
 import { CardLayoutComponent } from './card-layout/card-layout.component';
-import { ContactComponent } from './contact/contact.component';
-import { DangerousMaterialComponent } from './dangerous-material/dangerous-material.component';
+import { ContactComponent } from '../intervention-building/contact/contact.component';
+import { DangerousMaterialComponent } from '../intervention-building/dangerous-material/dangerous-material.component';
 import { FireHydrantComponent } from './fire-hydrant/fire-hydrant.component';
 import { FireProtectionComponent } from './fire-protection/fire-protection.component';
 import { ImplantationPlanComponent } from './implantation-plan/implantation-plan.component';
 import { ParticularRiskComponent } from './particular-risk/particular-risk.component';
-import { PersonRequiringAssistanceComponent } from './person-requiring-assistance/person-requiring-assistance.component';
+import { PersonRequiringAssistanceComponent } from '../intervention-building/person-requiring-assistance/person-requiring-assistance.component';
 import { CardLayoutFactoryDirective } from './shared/control-factory.directive';
 import { BuildingContactService } from './shared/services/building-contact.service';
 import {BuildingHazardousMaterialService} from './shared/services/building-hazardous-material.service';
