+++ src/app/intervention-survey/intervention-survey.module.ts
@@ -4,9 +4,9 @@ import { SharedModule } from '../shared/shared.module';
 import { BuildingComponent } from './building/building.component';
 import { BuildingDetailComponent } from './building-detail/building-detail.component';
 import { ImplantationPlanComponent } from './implantation-plan/implantation-plan.component';
-import { ContactComponent } from './contact/contact.component';
-import { PersonRequiringAssistanceComponent } from './person-requiring-assistance/person-requiring-assistance.component';
-import { DangerousMaterialComponent } from './dangerous-material/dangerous-material.component';
+import { ContactComponent } from '../intervention-building/contact/contact.component';
+import { PersonRequiringAssistanceComponent } from '../intervention-building/person-requiring-assistance/person-requiring-assistance.component';
+import { DangerousMaterialComponent } from '../intervention-building/dangerous-material/dangerous-material.component';
 import { FireProtectionComponent } from './fire-protection/fire-protection.component';
 import { FireHydrantComponent } from './fire-hydrant/fire-hydrant.component';
 import { ParticularRiskComponent } from './particular-risk/particular-risk.component';
@@ -54,16 +54,11 @@ import {CourseLaneComponent} from './course-lane/course-lane.component';
     FireHydrantComponent,
   ],
   providers: [
-    BuildingContactService,
-    BuildingHazardousMaterialService,
-    HazardousMaterialService,
     UnitOfMeasureService,
     FireHydrantTypeService,
     LocationTypeService,
     InterventionPlanFireHydrantService,
     LaneService,
-    PersonRequiringAssistanceTypeService,
-    BuildingPersonRequiringAssistanceService,
     PictureService,
     InterventionPlanService,
     InterventionPlanCourseService
